Added board class
Added the board class and the following methods:
isValidHeapChoice
isValidTokenChoice
removeTokens
checkForLoss
getTokensInHeap


Also added documentation for what each of the mothods does as well as comments that clarify specific steps in the process of some methods.

Changed the parameters of "isValidTokenChoice" to also take in the selected heap to assist in validation.